switch to vite and upgrade cypress
