+++ cypress/integration/spec.js
@@ -1,6 +1,6 @@
 describe('My First Test', () => {
   it('Does not do much!', () => {
-    cy.visit('http://localhost:4000')
+    cy.visit('http://localhost:3000')
     cy.wait(1000)
     cy.window().then(({ pages }) => {
       expect([...pages]).to.deep.equal([1, 2, 3])
