+++ example/vite.config.ts
@@ -1,0 +1,9 @@
+import vue from '@vitejs/plugin-vue'
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  plugins: [vue()],
+  esbuild: {
+    jsxFactory: 'h'
+  }
+})
