+++ package.json
@@ -14,7 +14,7 @@
     "lint": "xo",
     "build": "bili --format esm,cjs,esm-min --minimal",
     "prepublishOnly": "npm run build",
-    "example": "poi example/index.js --serve --jsx vue",
+    "example": "vite example",
     "commit": "git-cz",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
@@ -27,11 +27,12 @@
   "author": "egoist<0x142857@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "@vitejs/plugin-vue": "^1.4.0",
     "@vue/compiler-sfc": "^3.1.5",
     "babel-eslint": "^10.0.1",
     "bili": "^4.0.9",
     "commitizen": "^3.0.5",
-    "cypress": "^3.3.2",
+    "cypress": "^8.2.0",
     "cz-conventional-changelog": "^2.1.0",
     "eslint-config-prettier": "^3.3.0",
     "eslint-config-rem": "^4.0.0",
@@ -41,11 +42,10 @@
     "husky": "^1.0.0-rc.13",
     "lint-staged": "^7.2.0",
     "npm-run-all": "^4.1.5",
-    "poi": "^12.4.6",
     "prettier": "^1.15.2",
     "semantic-release": "^17.3.0",
+    "vite": "^2.4.4",
     "vue": "^3.1.5",
-    "vue-loader": "^16.5.0",
     "vue-router": "^4.0.10",
     "xo": "^0.23.0"
   },
